rag_chatbot-main/requirements.txt
# --- Web Framework & Server ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0 # ADDED: Production WSGI server for Vercel
pydantic==2.8.2 # Recommended explicit dependency

# --- Core LangChain and AI Suite ---
langchain==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.0.4
google-generativeai==0.8.5
langchain_community - 0.3.27

# --- Vector Database and Connectors ---
langchain-pinecone==0.2.0
pinecone-client==5.0.1
faiss-cpu==1.8.0


# --- Data Handling ---
pandas==2.2.2
numpy==1.26.4

# --- Utilities ---
pypdf==4.2.0
python-dotenv==1.1.1

boto3==1.34.150 


